PRINT FLOW LIST start
Many minor changes

+++ dhr_mark4/string_handling.py
@@ -35,7 +35,7 @@ def remove_useless(string) :
 	return(return_string)
 
 def string_to_int(string) : 
-	return float(string)
+	return int(string)
 	# num = 0
 	# for i in range(len(string)) : 
 	# 	num += (10**(len(string)-i-1))*int(string[i])
